django/requirements.txt
Django==3.1.6
psycopg2-binary==2.8.6
djangorestframework==3.12.2
django-filter==2.4.0